refactor: moved inventory modeld to dedicated package bumped version to 1.98.0 updated changelog
