+++ CHANGELOG.md
@@ -1,6 +1,16 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## [1.98.0]
+
+### Added
+- Introduced support for Inventory features:
+    - `Barcode` model
+    - `TagInfo` model
+    - `RecordInfo` model
+    - `SharedWith` model
+- Added the ability to represent sharing mode, barcodes, and tags for Inventory.
+
 ## [1.97.0]
 
 - Support newly added audit (activity) actions: DUPLICATE and RENAME
